chore: only allow releasing from release branches
This changes the release so that only release branches in the form of release/v0.y are accepted.

It changes the release Branch creation into a "Release Branch Cutoff". This leads to us only having one point in which the version file is bumped for y-streams (0.17,0.18. etc.).

During the release this file is touched up only to remove the -dev suffix after which the tag is pushed.

For z-streams (0.17.z) a new commit with a bump is still needed this is not included yet.